Update README for steps on deploying and connecting with Metamask
- Update README to show steps for deploying tokens and connecting with Metamask
- Add public and private keys to `.env` file
- Update tests to use typescript
