+++ hardhat/.env
@@ -1,0 +1,6 @@
+ALICE_KEY=0x784e2c4D95c9Be66Cb0B9cda5b39d72e7630bCa8
+ALICE_PRIV_KEY=5094f257d3462083bcbc02c61d98c038cfa71cdd497834c5f38cd75010ddb7a5
+BOB_KEY=0xF4d5B303A15b04D7C6b7510b24c62D393805B8d7
+BOB_PRIV_KEY=78785c4ab4ba44b83509296af86af56ff00db79ba26a292d0556a4b4e8cea87c
+CHARLIE_KEY=0x67C94d4a4fab02697513e4611A4742a98879aD56
+CHARLIE_PRIV_KEY=417fbb670417375f2916a4b0110dc7d68d81ea15aad3e6eb69f166b5bed6503f
